Add files via upload
The plain_text file has the original info you'd like to decrypt this can be either an image or a message, it is then decrypted and put into cipher text and can then be encrypted and saved to plain_text1. This is all done using the key in the key file.